example config: Refine template for BlueSky
Re-jig the template to be a little more concise in the first 250
characters to handle limitations of other federated social media
platforms like BlueSky (AT protocol).

This moves a subset of all the tags to the first line so that there's a
good chance critical ones will make it across bridges like Bridgy Fed.